Home Screen

+++ app/src/main/java/com/example/cryptoapp/data/remote/dto/CoinTickerDto.kt
@@ -8,20 +8,20 @@ data class CoinTickerDto(
     @SerializedName("beta_value")
     val betaValue: Double,
     @SerializedName("circulating_supply")
-    val circulatingSupply: Int,
+    val circulatingSupply: Long,
     @SerializedName("first_data_at")
     val firstDataAt: String,
     val id: String,
     @SerializedName("last_updated")
     val lastUpdated: String,
     @SerializedName("max_supply")
-    val maxSupply: Int,
+    val maxSupply: Long,
     val name: String,
     val quotes: TickersQuotes,
     val rank: Int,
     val symbol: String,
     @SerializedName("total_supply")
-    val totalSupply: Int
+    val totalSupply: Long
 ){
     fun toCoinTicker(): CoinTicker{
         return CoinTicker(
